refactor asset_lib to sahlib
